move device -> Device/device
